Abduulah and Ayman

+++ restaurants/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect
+
 from .models import Restaurant, Item
 from .forms import RestaurantForm, ItemForm, SignupForm, SigninForm
 from django.contrib.auth import login, authenticate, logout
@@ -58,7 +59,13 @@ def restaurant_detail(request, restaurant_id):
     }
     return render(request, 'detail.html', context)
 
+def no_access(request):
+    return redirect('no-access')
+
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect('signin')
+
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -73,8 +80,15 @@ def restaurant_create(request):
     return render(request, 'create.html', context)
 
 def item_create(request, restaurant_id):
-    form = ItemForm()
+    if request.user.is_anonymous:
+        return redirect('no-access')
+
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if not(request.user.is_staff or restaurant.owner == request.user):
+        return redirect('no-access')
+
+    form = ItemForm()
+    
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -89,7 +103,11 @@ def item_create(request, restaurant_id):
     return render(request, 'item_create.html', context)
 
 def restaurant_update(request, restaurant_id):
+    if request.user.is_anonymous:
+        return redirect('no-access')
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not(request.user.is_staff or restaurant_obj.owner == request.user):
+        return redirect('no-access')
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
@@ -103,6 +121,8 @@ def restaurant_update(request, restaurant_id):
     return render(request, 'update.html', context)
 
 def restaurant_delete(request, restaurant_id):
+    if request.user.is_anonymous:
+        return redirect('no-access')
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
